# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillTwin AI Mentor API",
    "description": "Contract and intelligence source-of-truth for the SkillTwin personalized AI mentor platform. Exposes adaptive goals, winding journey roadmaps, evidence-based learner twin state, real-time mentor guidance, and spaced retention retrieval practice.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/teach/transcribe": {
      "post": {
        "tags": [
          "Teach Mode"
        ],
        "summary": "Transcribe Learner Voice Explanation (Speech-to-Text)",
        "description": "Transcribes learner speech into text for Feynman teach-back analysis.\nValidates audio format, duration, and extracts operational transcription.",
        "operationId": "transcribe_audio_explanation_api_v1_teach_transcribe_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transcribe_audio_explanation_api_v1_teach_transcribe_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioTranscriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teach/evaluate": {
      "post": {
        "tags": [
          "Teach Mode"
        ],
        "summary": "Evaluate Teach-Back Explanation & Update Learner Model",
        "description": "Evaluates a learner's teach-back explanation using the Feynman technique.\n\nPipeline:\nTranscript + Learner Context -> LLM Evaluator -> EvaluationService -> LearnerService -> State & Journey Recalibration.\n\nEnforces mandatory separation: The LLM evaluator does not directly write to the database.\nOutputs are validated by EvaluationService and deterministically persisted by LearnerService.",
        "operationId": "evaluate_teach_explanation_api_v1_teach_evaluate_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeachEvaluateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeachEvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AudioTranscriptionResponse": {
        "properties": {
          "transcript": {
            "type": "string",
            "title": "Transcript"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "detected_language": {
            "type": "string",
            "title": "Detected Language",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "transcript",
          "duration_seconds",
          "word_count",
          "confidence"
        ],
        "title": "AudioTranscriptionResponse"
      },
      "Body_transcribe_audio_explanation_api_v1_teach_transcribe_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Recorded audio explanation (WAV, MP3, M4A, WebM, OGG)"
          },
          "concept_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Id",
            "description": "Optional concept context hint for acoustic modeling"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_transcribe_audio_explanation_api_v1_teach_transcribe_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TeachEvaluateRequest": {
        "properties": {
          "concept_id": {
            "type": "string",
            "title": "Concept Id"
          },
          "explanation_text": {
            "type": "string",
            "minLength": 5,
            "title": "Explanation Text",
            "description": "Learner explanation text transcribed or typed"
          },
          "audio_duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Duration Seconds"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "self_reported_confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Self Reported Confidence"
          }
        },
        "type": "object",
        "required": [
          "concept_id",
          "explanation_text"
        ],
        "title": "TeachEvaluateRequest"
      },
      "TeachEvaluationResponse": {
        "properties": {
          "concept_id": {
            "type": "string",
            "title": "Concept Id"
          },
          "conceptual_accuracy": {
            "type": "integer",
            "title": "Conceptual Accuracy"
          },
          "completeness": {
            "type": "integer",
            "title": "Completeness"
          },
          "reasoning": {
            "type": "integer",
            "title": "Reasoning"
          },
          "confidence": {
            "type": "integer",
            "title": "Confidence"
          },
          "transfer": {
            "type": "integer",
            "title": "Transfer"
          },
          "misconceptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Misconceptions"
          },
          "missing_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Concepts"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          },
          "evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Id"
          },
          "updated_mastery_score": {
            "type": "number",
            "title": "Updated Mastery Score"
          },
          "updated_confidence_score": {
            "type": "number",
            "title": "Updated Confidence Score"
          },
          "updated_status": {
            "type": "string",
            "title": "Updated Status"
          },
          "feedback_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Summary"
          },
          "key_strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Strengths"
          },
          "growth_areas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Growth Areas"
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluated At"
          }
        },
        "type": "object",
        "required": [
          "concept_id",
          "conceptual_accuracy",
          "completeness",
          "reasoning",
          "confidence",
          "transfer",
          "recommendation",
          "updated_mastery_score",
          "updated_confidence_score",
          "updated_status",
          "evaluated_at"
        ],
        "title": "TeachEvaluationResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}